requirements.txt
colorlog==4.0.2
pysam==0.15.3
pytabix==0.1
rstr==2.2.6
